requirements.txt
flask_dropzone
cx_Oracle
sqlalchemy
Unidecode
flask
flask-api
python-box
psutil
requests
pandas